Refaktor: InMemory- és JpaUserService közös elemek kiemelése
